src: main: Add brute force implementation of single page processing
Processing of representative patches is done in chunks. As the lore API
responds with an Atom feed with maximum 200 patches, we need to repeat
the steps of:

  1. Request to lore API an Atom feed with patches from a target mailing
     list from a specific chunk;
  2. Process the lore API response for a sequence of representative
     patches;
  3. If there are more representative patches that need to be processed,
     update chunk to the adjacent one and repeat steps 1 to 3.

To take care specific of step 2, i.e., processing of a lore API response
into a sequence of representative patches, implement brute force logic
that showcases a processing of a pre-fetched chunk from the `amd-gfx`
list, and prints it to the standard output.

Note.: Also, introduce `lore-peek` binary crate root `main.rs`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>