src: lib: Extract Lore API response patch processing to lib modules
The processing of a Lore API response into a vector of representative
patches ordered from latest to oldest is completely done inside the main
method of the `lore-peek` binary crate root (i.e., the file
`src/main.rs`). This is bad for a several number of reasons, like:
scalability, encapsulation, testing, etc.

To solve this, extract all the necessary data structures corresponding
to the Lore API response entities and their related methods into
library modules. Namely, two modules were extracted: `lore_response`,
which only contains the definition of the type `LoreResponse`, and
`patch`, which contains the definition of the type `Patch`, its subtypes
`Author` and `MessageID`, and its methods for representative patches
processing.

Note that the structures related to a "Lore session", which correspond
to a sequence of processing of Lore API responses to (potentially)
navigate all the patch history, aren't included into this extraction.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>