src: lore_session: Extract remaining patch processing to `LoreSession`
The crate root of `lore_peek` main binary crate (i.e., `src/main.rs`)
still contains a lot of logic in terms of a "Lore session". We define
a "Lore session" as a vector of representative patches ordered from
latest to oldest, a map of all processed patches, and it represents the
state of the fetching of patches to a target mailing list.

In this sense, extract the remaining logic from the main binary crate to
a dedicate module.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>