src: patch: Refactor `LoreResponse` into `PatchFeed`
The type `LoreResponse`, which actually represents a vector of the type
`Patch`, is badly named, as it doesn't relate to the response from Lore,
but actually the Atom feed returned when the GET request is successful.

In this sense, rename the type to `PatchFeed`, remove the dedicated
module `lore_response`, and move the declarations and definitions to the
`patch` module.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>