src: patch: Fix regex to capture patch tag from patch title
The capturing a patch tag from a patch title is crucial to a correct
parsing of a patch metadata, and (consequently) crucial to the
processing of representative patches, which is one of the core features
of `lore-peek`. The regex responsible for this is defined as a field of
`PatchRegex` with the name of `re_patch_tag`.

The regex correctly handles situations in which the title starts with
the patch tag. However, when faced with a title, say

```
  [GSoC][PATCH X/Y] foo: bar: Do foo-bar
```

the regex will capture `[GSoC]`, which indicates that it really just
captures the first string surrounded by square brackets.

In this sense, fix the regex for capturing the patch tag from a patch
title.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>