src: main: Implement dynamic single request to Lore with brute force
Currently, `lore_peek` uses a pre-fetched Lore response as the input for
the processing of representative patches.

In this context, implement a dynamic single HTTP request to Lore API to
fetch an Atom feed containing the last patches sent to a target list
that is passed as argument. The request is made using `reqwest`, and the
logic handles the three necessary cases:

  1. 200 OK returning a non-empty Atom feed (success case);
  2. 200 OK returning an empty Atom feed (normal fail case);
  3. Not 200 OK (abnormal fail case).

This is implemented completely in the crate root of the main binary
crate, which points to the brute force used. In the future we should
extract this logic to a dedicated type, and maybe a dedicated module.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>