src: lore_api_client: Extract Lore request logic to `LoreAPIClient` type
All logic related to building a Lore API request for a patch feed,
making the request, and handling errors, is done directly into the crate
root of the main binary crate of `lore_peek`. This isn't versatile and
won't scale well if we want to make use of other Lore API requests, like
advanced searches and raw patch fetching.

To mend these issues, extract all the logic into the `LogicAPIClient`
type defined in the `lore_api_client` module. This type doesn't have any
fields, and only one associated function to handle the request for
a patch feed from a target mailing list. It follows that future
categories of Lore API requests should be implemented as associated
functions like this, but the implementation doesn't deprives us to build
a more robust type that will have complex instances, and so on.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>