src: main: Implement brute force pagination to Lore request for patch feed
Because Lore limits the response to a request for patch feed to 200
patches, in other words, it paginates the *response* for a given
request, we need to implement a mechanism that can (potentially) fetch
and process all the pages of a response.

To accomplish this, implement a brute force logic inside the crate root
for the main binary crate to iterate through the pages and process
patches until it processes, at least, the number of representative
patches required by the user.

The implementation manipulates the paramater `o` of the Lore API, which,
for patch feed requests, indicates the minimum exclusive index of the
first patch from the response. More specifically, it repeats the process
of fetching a page, processing the patches from the page, updating the
representative patches, and adding 200 to the value of `o`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>