src: lore_session: Extract pagination Lore request for patch feed into `LoreSession` method
Extract all the logic pertaining to process the paginated responses of
a Lore request for the patch feed of a target list to the method
`process_n_representative_patches` of the `LoreSession` type.

This extraction made possible the encapsulation of practically all
implementation details and dependencies from the crate root of the main
binary crate. More specifically, the only type needed for a caller that
wishes to fetch a number of representative patches is the `LoreSession`,
and the `serde_xml_rs` dependency (responsible for deserializing the
patch feed) is encapsulated into the `lore_session` module. Also, the
public methods `process_patches` and `update_representative_patches` are
turned into private, as they are now only called by the extracted
method.

The only downside is that now we need to incorporate the values of `o`,
i.e., the `min_index`, and of the target list into the `LoreSession`
type.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>