src: lore_session: Refine error handling of failed Lore feed request
When requesting a page from the patch feed of a target list to Lore,
there can be 3 fail scenarios (represented by the `FailedFeedRequest`
enum):

  1. An unknown error;
  2. A response status code that is non 200 Ok;
  3. An empty feed.

For each, the caller may want to handle differently. However, when
processing a number of representative patches, the `LoreSession` type
panics without giving the caller a chance to handle it.

In this sense, refine the error handling of a failed Lore API request
for a patch feed, by relaying this errors to the caller of the method
`process_n_representative_patches`. Note that the tuple variant
`StatusNotOk` had its `u16` switched to `reqwest::blocking::Response`
for better debugging info.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>