src: lore_session: Add pagination of representative patches for front-end
As the `LoreSession` type represents a session in which a target patch
feed is fetched, processed for representative patches, and stores them
ordered from latest to oldest (represented by the attribute
`representative_patches_ids)`, we need a way for clients, like
a front-end component, to have an easy access to a page of these. Note
that the concept of "page", in the context of the front-end, is
different from the one in the back-end, i.e., a page of the *response*
to a patch feed request.

With this in mind, implement the public method
`LoreSession::get_patch_feed_page` that, given a page size and a page
number, returns a `Vec<&Patch>` for easy access to iterate over the page
in order.

Note that we may need to revise if it is more performatic to return
a slice of the field `representative_patches_ids`, and directly query
the `LoreSession` type for an immutable reference to the `Patch` type
with `get_processed_patch`. But, "early optimization is the root of all
evil", so lets cross this bridge when/if we need.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>