src: main: Introduce proof of concept of front-end client
Just using the model laid out by the previous commits, introduce
a simple terminal front-end (view) client that uses the model
represented by the `LoreSession` type to display the feed of
representative patches in pages. This effectively turns the main binary
crate into the first draft of the `lore_peek` app.

In this starting draft of an app, the user can consult the sequence of
representative patches of the patch feed from a target mailing list.
This sequence is partitioned in same sized pages, which the user can
navigate to older (forward) and to newest (backwards) pages at runtime,
and gracefully quit the application. At last, the user defines the
target list and page size by passing CLI arguments. The processing of
older pages happens on demand, and pages already processed don't get
processed again.

Note that we aim to use a real TUI framework in the future. Right now,
I have in mind the `ratatui.rs` crate, as it has a great look and feel,
and seems to be versatile and powerful enough for our needs, while still
being lightweight.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>