src: lore_session: Fix out-of-bounds bug when calling `get_patch_feed_page`
The method `LoreSession::get_patch_feed_page` has a bug in which,
depending on the size of `representative_patches_ids`, and the arguments
of `page_size` and `page_number`, the function could panic due to not
handling cases in which the requested page is out-of-bounds. This is
due to the fact that the mailing lists patch feeds are finite, so it is
possible to reach the end of the feed, in which subsequent (front-end)
page request will cause this problem.

To fix this, check if the lower and upper ends of the page requested are
out-of-bounds, and return `None` or cap the upper end at the max index of
`representative_patches_ids`, in case the size of the vector is smaller
than the lower end or smaller than the upper end, respectively.

To accompany this change, also adapt the draft `lore_peek` front-end to
handle these edge cases accordingly.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>