src: lore_session: tests: Add test suite for `LoreSession`
To continue expanding the tested module in the project, add a test suite
for the `LoreSession` type, the main type of the
`lore_peek::lore_session` module. This commit only adds a simple
initialization test, as the core of the fetching will be tested on
a future commit, because we need to refactor `LoreAPIClient` to be able
to mock the HTTP request.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>