src: lore_api_client: Decouple Lore API client from `LoreSession` w/ trait `PatchFeedRequest`
In the context of fetching a patch feed from a target mailing list, all
the logic related with Lore API request is contained in the concrete
type `LoreAPIClient`. This type is heavily coupled with `LoreSession` in
the sense that it is instantiated and used inside the
`process_n_representative_patches` method (the core of the process).

To decouple the type from `LoreSession` for more flexibility *and*
testability, create the trait `PatchFeedRequest` and the type
`BlockingLoreAPIClient` (which is the same as `LoreAPIClient`, but
evidentiates that it is a blocking one).

By using trait bounds on `PatchFeedRequest` (and future API features),
we can both mock the HTTP client, and have more choices of clients for
the real implementation.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>