src: lore_session: tests: Implement tests for core of processing representative patches
The `LoreSession` type encapsulate all the logic for fetching and
processing a patch feed for representative patches, which is the most
critical feature of `lore-peek`. On the other hand, there was no unit
test for the method responsible for this important task, i.e., the
`process_n_representative_patches` method.

In this context, add two test functions to test this core unit of
`lore-peek`. It is important to note that these tests mock the Lore API
request for a patch feed to avoid flakyness.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>