Cargo.toml: Add `ratatui` and `color-eyre` dependencies
As the focus now is to start integrating the `ratatui` library for
building TUIs into the project, add the dependency `ratatui` to the
project.

As a collateral effect, also add `color-eyre` to generate hooks for
panics and error handling. When using `ratatui` we need to save the
current terminal buffer, change to an alternate screen, enable
"raw-mode" (which turns off the normal input and output processing by
the terminal), and (after the application has finished) restore the
terminal. If the application panics in the middle of execution, without
custom panic hooks for restoring the terminal state, the terminal
session will be messed up.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>