src: main: Implement dirty prototype for `lore-peek`
Considering that `lore-peek` aims to be an application that is more than
just a TUI to the flow of patches from the mailing lists archived at
`lore.kernel.org`, it is undeniable that the core of the application (in
terms of the most critical part) resides in this task. The "back-end" of
this task has already been implemented in the `lore-peek` lib, so the
next step to produce a prototype for the full application is to
integrate a "front-end", which will be implemented using "ratatui".

In this context, implement a dirty prototype for the full `lore-peek`
application. The prototype consists of a TUI that starts on a screen
asking for the user to input the target mailing list. and another that
list the latest patchsets from that mailing list, allowing the user to
navigate the full history with pagination.

The term "dirty" is used because of four reasons:

  1. Most of the code for the Controller and View components of the
     application can (and *should*) be better encapsulated.
  2. There are no tests.
  3. There is no handling of recoverable errors.
  4. The thread blocks when the back-end processes for patchsets,

Tackling the above points should be the focus in the next cycle of
development.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>