Cargo.toml: Update package name to `lore-peek`
The project's package name is `lore_peek`, in snake_case, but we want it
to be in kebab-case, i.e., `lore-peek`. The package name was in
snake_case because I thought the it needed to be that way as package
imports are done using snake_case.

However, as the import syntax defaults to snake_case, despite the
package name being on kebab-case, enforce the latter.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>