src: ui: Use `ListState` for list item selection in latest patchsets
The selection of an item in the latest patchsets screen was implemented
without leveraging the `StatefulWidget` trait from `ratatui`, which are
widgets that can also encapsulate some state.

The `List` widget used for listing the latest patchsets from a target
mailing list is one of those widgets, which allow us to set the offset
of the list and the selected item, while also providing built-in methods
to style the highligthing.

In this sense, use the type `ListState` for the selection functionality
along with the function `render_stateful_widget`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>