src: app: Add 'Patchset Details and Actions' screen
[Why]

One of the core features `lore-peek` proposes to offer is the display of
information on the patchsets (i.e., a sequence of correlated patches)
that flow through the mailing list, and also allow users to run actions
on those, like: bookmarking for later consulting, applying the patchset
to a target git tree, building/deploying a kernel based on that, etc.

[What]

To provide this, add a screen named `Patchset Details and Actions` that,
currently, displays metadata about the patchset, as well as the
contents of every patch (and the cover letter) of the series. The layout
implemented has three panes: the right half with the contents of the
series, the top left with the patchset metadata, and the bottom left
which shall contain the list of assignable actions on the patchset. As
each action may involve a lot of logic, the plan is to add them
incrementally (probably starting with the bookmark action).

[How]

The core logic was implemented in the `src/lore_session` module with the
functions `download_patchset` and `split_patchset`, which handle the
downloading and splitting of the mbox file into individual patches for
displaying, respectively. The first task is done by the `b4` tool, which
`lore-peek` assumes the system has installed.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>