src: lore_session: Handle case for patch not ending on "--" in `extract_patch`
When splitting `.mbx` files containing multiple patches (done by
`split_patchset` and `extract_patch` functions), the logic must rely on
many patterns and heuristics. One of the patterns that it depends on is
the presence of a double dash line ("--") followed by a git version,
which (I assume) is necessary for the patch to be applicable. In recent
tests, I discovered that this case occurs, causing multiple patches to
be interpreted as one by our code. I've concluded that, indeed, this
isn't `b4` mangling the file, as the patch was sent that way[1], which
makes me think that it is, in fact, a "wrong" edge case.

However, as this bug in the logic relates only to displaying the patch
for visualization, handling this case won't affect the correctness of
our implementation. On the contrary, it would make it more friendly to
new contributors and probably more robust.

With this context, make `extrac_patch` handle this case by relying on
a fallback pattern of the line before "Subject:" (the pattern used to
determine the start of a patch) being "From git@z Thu Jan  1 00:00:00
1970".

[1] https://lore.kernel.org/git/20240715003519.2671385-2-e@80x24.org/

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>