src: app: Add bookmarking actions
The "Patchset Details and Actions" screen is planned to house a great deal
of the functionality of `lore-peek`, including the (un)bookmarking of
patchsets for later consulting.

In this sense, add the bookmarking and unbookmarking of patchsets via
the "Patchset Details and Actions" screen, together with a similar
screen to the "Latest Patchsets" screen, i.e., a list of patchsets with
same style (for now), that is named "Bookmarked Patchsets".

As the data related to bookmarked patchsets is supposed to be
persistent, we introduce (probably the first) JSON file that shall
reside in directories like `XDG_DATA_HOME/lore-peek`, with
`XDG_DATA_HOME` usually being `HOME/.local/share`. Currently, this
location represented by the constant `BOOKMARKED_PATCHSETS_PATH` is
hardcoded to my specific fs, so we need to generalize this.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>