src: app: Make "Mailing List Selection" list stateful and interactive
The last commit implemented a dynamic list of the available mailing
lists available in Lore. This helped a lot, as now the doesn't need to
know by heart the specific list IDs on lore. However, the set of
available list shown only serves as a guideline, because the user has to
correctly type the list name either way.

To ease even more this process, add state for the set of available lists
by highlighting an specific item from the list that is selectable with
the up and down arrow keys, along with completion with the Tab key.

There are some trade-offs to this statefulness and interactiveness:

  1. This screen uses O(n) operations on the size of the available
     mailing list for each time the user types or deletes characters,
     but considering that the number of list archived on Lore is
     "constant" and small, this probably isn't a problem.
  2. Remapping of Tab for bookmarked patchsets, as well as the limit of
     arrow keys for navigating through the items of the set of lists (we
     can't use "j" or "k", but this is possible).

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>