Rename project from `lore-peek` to `patch-hub`
The project calle so far as `lore-peek` was born as a rework of the
kworkflow feature `kw patch-hub`. Now that the project has matured
a little, we are incorporating it to the kworkflow organization hosted
at GitHub and are planning (maybe as the immediate next step) to
integrate it to the `kw` toolset.

As we are almost sure that this will be the way for the `kw patch-hub`
feature, rename the project from `lore-peek` to `patch-hub`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>