src: lore_session: Implement fetching and processing of available lists
Lore API provides requests to obtain the available mailing lists (along
with descriptions), which have an archive on `lore.kernel.org`. This is
accomplished by sending HTTP GET requests to the base URL, which accepts
a query parameter of `o` that represents the pagination of responses,
just like with patch feeds.

To avoid requiring the user to know by heart the exact mailing lists
names he wishes to consult, and also provide them with a description of
the list, implement the "back-end" of this process.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>