src: app: Implement dynamic "Mailing List Selection" screen
Currently, the screen in which the user can input the target mailing
list that he/she wishes to consult doesn't assist the user in any way,
forcing the memorization of the exact string that IDs the list in Lore.

To enhance the functionality of this screen, implement the dynamic
fetching of the available mailing list in Lore, along with the sorting
and live displaying the possible lists based on the buffered user input
(i.e., if the user has a prompt with "b4", the "b4-sent" and "b43-dev"
will be the lists shown for completion).

Note: This commit only implements the "front-end" of the task of
fetching, processing and displaying the lists, as the "back-end" was
done by the previous commit.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>