src: app: Encapsulate state of "Mailing List Selection"
The "Mailing List Selection" screen is getting more complex, and we want
to implement autofill and a stateful list for the available lists.

Considering that the values that configure the state of this screen are
root fields in `App` (currently, `mailing_lists` and `target_list`), and
that the other screens are encapsulated in a `<ScreenName>State` type,
extract this values into a single organized type
`MailingListSelectionState` and make the necessary adjustments.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>