src: app: Handle invalid target list for "Mailing List Selection"
If the user hits enter with an invalid target list in the "Mailing List
Selection" screen, i.e., submits the inputted target list that is either
colored cyan or red, the program will crash, because it will attempt to
fetch for a patch feed from an invalid list from Lore API, which, in
turn, will panic.

To avoid this, only attempt to request for the patch feed from valid
lists by checking if the current target list corresponds exactly to one
of the valid mailing lists. This virtually makes the process infalible
from the perspective of generating panics from invalid requests.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>