src: app: config: Make configurations based on user's env
Currently, some configurations options like the page size, the cache
paths, and so on, are hardcoded and actually point to my specific
filesystem structure.

To move the feature from a controlled prototype into a prototype that is
testable by a broader number of users, make the configurations dynamic
(i.e., determined at runtime) and based on environments variables. This
represents the first step of the integration of `patch-hub` into `kw`,
the second being developing code on `kw` side that sets the necessary
environment variables and runs `patch-hub`.

Note: The implementation should be revised in the future, as `Config` is
a "data type" of the `App` type (represents the app state), but some
configs are being passed around a lot, like to the UI component and
other `App` fields (substates of the app).

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>