src: app: Make (un)bookmarking a patchset a transaction
When we bookmark a patchset, this action is only persistent in the file
`bookmarked_patchsets.json` when the application gracefully exits. This
opens margin for panics that may cause this file to become inconsistent.

In this sense, make the (un)bookmarking of a patchset a transaction by
consolidating it in the persistent file each time a patchset is
(un)bookmarked, avoiding panics after the it to make this (degenerated)
database inconsistent.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>