src: lore_session: Add caching for `b4` downloads
When downloading an applicable patchset using the `b4` tool, there is no
check to see if the target file is already cached. This results in
unnecessary slowdowns, despite `b4` being smart enough to check for the
caching on their side, there still a slowdown.

To correct this, check to see if the target applicable patchset is
already cached before invoking `b4`, and avoid doing so if it is.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>