src: lore_session: Implement unit to generate patch reply template
Part of the task of implementing the action to reply a patch with
a `Reviewed-by` tag is to generate a base reply file similar to the ones
mail clients generate when you use the "reply" option.

In this sense, implement the unit responsible for this task. More
specifically, this unit generates a type `String` (for versatility) that
has the same `Subject:` header (with the `Re:` included) and headers
other than `From:`, `Date:`, and `Message-Id:`, along with the rest of
the message, i.e., its body, quote-replied.

It is important to notice that this unit will be useful not only for the
implementation of the `Reviewed-by` tag, but for any action on a patch
that involves replying to it.

Helps: #13

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>