src: lore_session: Implement unit to extract git reply command from patch HTML
For any single patch (actually, any message) hosted in Lore, if accessed
by using a GET HTTP request like

  https://lore.kernel.org/<list>/<message-ID>/

the response payload is a HTML file that contains, besides other
contents (the actual patch, the thread overview, and so on), a template
of a `git send-email` that can be used to reply the patch to the
interesting parties.

In the context of adding the action of replying patches, implement
a unit that parses this HTML and builds a `Command` type that represents
this base `git send-email` reply command.

Note.: To be safe about not spamming the mailing lists, add the
`--dry-run` flag by default. At this point, there is no code in which
sending a patch is possible, but enforce this until we correctly
validate the full feature.

Helps: #13

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>