src: lore_api_client: Implement request to patch HTML
Every patch archived in Lore has an HTML version that is accessible by
making a GET HTTP request to

  https://lore.kernel.org/<target-list>/<message-id>/

This HTML version of the patch has additional informations that are not
present in the raw version of the patch, like a reply template for
a `git send-email` command.

In this sense, add the `PatchHTMLRequest` trait for requesting this
version of patches and implement it for `BlockingLoreAPIClient`.

Note: The implementation of this trait ended up quite similar to the one
for `AvailableListsRequest`, differing only in the endpoint requested,
so maybe we need to make a more general trait for these cases.

Helps: #13

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>