src: lore_session: Add mediator to reply to collection of patches w/ `Reviewed-by`
The last commits implemented unit by unit of the process of replying
a patch with `Reviewed-by` tag. However, the project lacks a mediator to
tie every unit to handle replying to set of patches (not necessarily
from the same patchset).

In this context, implement the `prepare_reply_patchset_with_reviewed_by`
function that, for each patch in a vector passed as argument, creates
a reply template, adds the `Reviewed-by` tag (with the signature passed
as argument), and prepares a vector of `Command` that represent the
sequence of `git send-email` commands needed to actually send the
replies.

It is important to stress that this mediator doesn't commit to the
actual sending of patches, leavig this responsability to the caller.
Another point worth noting is that we may want to make this mediator
leaner in the future (there are too many args being passed to it).

Helps: #13

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>