src: app: Add action `reviewed-by` to "Patchset Details and Actions"
The (arguably) main screen of `patch-hub` is the "Patchset Details and
Actions" screen, in which the user can consult the patchset metadata,
individual patch contents, and run actions on these patches. Currently,
the only action available is the bookmark/unbookmark action, which
adds/removes the patchset to a local database for later consultation.

To finalize the last commit sequence, add the `reviewed-by` action to
this screen, that allows the user to reply to every patch in the series
with a `Reviewed-by` tag.

Note: The `git send-email` commands are all run with the `--dry-run`
option, which makes everything *but* sending the email. When we are
comfortable with the way things are, we need to remove this option (or
make it optional) for this added feature to actually be useful.

Closes: #13

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>