src: app: Add visual mark to successful replies w/ `Reviewed-by` patches
Although the action to sent replies with the `Reviewed-by` tag is
implemented (although just with `--dry-run`), the application doesn't
track to which patches these actions were applied, neither if they were
successful or not.

To remedy this, implement a mark to the preview title of the patches in
the series which were successfully replied with the `Reviewed-by` tag.
Patches as such, will contain the preview title as "Preview [REVIEWED]",
instead of just "Preview". It is important to notice this type of
visualization is just a proof-of-concept, as the persistent tracking of
successfully replied patches can be used in other ways.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>