.github: workflows: release: Introduce CI pipeline for releases w/ `cargo dist`
In the Rust community, the `cargo dist` tool is used for quickly setting
up release workflows that integrates well with GitHub Actions.

To automate the process releasing portable binaries of `patch-hub` that
can be distributed through GitHub Releases, add a GitHub actions CI
pipeline, defined in `release.yml` that follows the usual `cargo dist`
workflow for cutting releases of your apps and distributing binaries
[1].

It is essential to note that this commit is an introduction for this CI
pipeline, as `release.yml` was generated by the `cargo dist init`, in
which the only configuration I made was to build for the target
platforms `x86_64-unknown-linux-gnu` and `x86_64-unknown-linux-musl`.

[1]: https://opensource.axo.dev/cargo-dist/book/introduction.html#plan-build-host-publish-announce
[2]: https://doc.rust-lang.org/rustc/platform-support.html

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>