Cargo.toml: Vendor `openssl` dependency
To handle HTTP requests, `patch-hub` uses the `reqwest` crate, which
depends on the `openssl` crate. There are two issues related to this
dependency:

  1. When running a pre-compiled release of `patch-hub`, the `openssl`
     lib fails to dynamically link, which resulting in the app crashing
     at start up (issue #19).
  2. When trying to build the `x86_64-unknown-linux-musl` trough the CI
     pipeline defined in `release.yml`, we encounter errors similar to
     the previous point (part of issue #18).

To fix both of these problems, vendor the `openssl` dependency by
explicitly declaring it as a dependency and enabling the `vendored`
feature. This is actually a fairly common practice in the Rust community
when dealing with this dependency (so much it has a plug-n-play flag for
it).

The obvious drawback, is that we are absorving this crate into our
codebase, which results in longer compile times (in my setup,
approximately doubled the time), and bloats our executables.

Closes: #19

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>