Cargo.toml: Remove unnecessary entry `packages_install`
When I was trying to add the release CI pipeline through `cargo dist`
for the `-musl` target, I tried solving it by adding an entry

  `packages_install = ["libssl-dev"]`

to (what I thought) was the matrix that `release.yml` used for extra
dependencies. This didn't solve the problem, but I forgot to remove it
and, as it did nothing, flew under the radar.

As such, remove this entry from `Cargo.toml`.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>