Cargo.toml: Add support for the `x86_64-unknown-linux-musl` target
For Linux platforms, Rust offers two generic targets: one that uses the
GNU libc (`x86_64-unknown-linux-gnu`) and the other that uses the musl
libc (`x86_64-unknown-linux-musl`). Because of problems with the
`openssl` dependency, we couldn't add support for the latter.

Now that we vendored the `openssl` depedency, this problem is solved, so
add support for the `x86_64-unknown-linux-musl` platform using `cargo
dist init`, which simply adds this new target to the `targets` value of
the `[workspace.metadata.dist]` configuration.

Closes: #18

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>