src: ui: Add `#[allow(dead_code)]` attribute to `centered_rect` func
When bootstrapping `patch-hub` integration with `ratatui.rs`, I used the
tutorial for the JSON editor [1] as a reference. This tutorial shows
a utility function named `centered_rect` for displaying centered pop-ups
and dialog boxes. This seemed useful, so I added it to the codebase.

The time to use it is taking too long, and like a flaky test shouldn't
remain in a test suite, unused code shouldn't stay in the codebase.
However, as this function will surely be handy in the future, add the
`#[allow(dead_code)]` attribute to suppress the warning caused.

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>