src: lore_session: Create cache dir before downloading patchset
When downloading a patchset with the `b4` tool, we use the function
`download_patchset` defined in the `lore_session` module, that accepts
an output directory as an argument that is considered a cache directory
for patchsets. However, this function doesn't check if this path exists,
which can result in ugly crashes.

Add check for path existence and create it, in case it doesn't.

Closes: #21

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>