Cargo.toml: Substitute vendoring of `openssl` by `rustls` feature
The dependency we use to make HTTP requests, `reqwest`, depends on
`openssl`. In the recent commit 565e65c, we had to solve portability and
release issues by vendoring the `openssl` crate as an explicit
dependency defined on `Cargo.toml`.

Because having vendored software shouldn't happen unless we have a solid
motive to substitute the vendored `openssl` dependency` by the `rustls`
feature of `reqwest`. We also added the flag `default_features = false`,
to (1) avoid adding unnecessary dependencies, and to (2) not have
`openssl` preceding `rustls`.

Note that `reqwest` may be overkill (`curl` could do it), so not using
it may be the most elegant alternative, but this solution is now enough.
It is also important to mention that we can't avoid using a TLS/SSL
implementation, as any raw HTTP requests to Lore API redirects to the
HTTPS scheme.

Helps: #22

Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>