src: main: Don't return "Bookmarked Patchsets" when there are no entries
When consulting bookmarked patchsets, if the patchset is the only
bookmarked one and the user unbookmarks it, the application returns to
an empty "Bookmarked Patchsets" screen. Even worse, if the user hits
ENTER, as if it were to select an highlighted entry, the application
panics.

To fix this strange behavior and the mentioned bug, return "Mailing
Lists Selection" screen when removing the last bookmarked patchset.

Closes: #12

[Maintainer edits]
- Change commit message to convey the context of the strange behavior
  and bug, and how it fixes both with a single solution.

Signed-off-by: Thiago Duvanel <thiago.duvanel@usp.br>
Reviewed-by: David Tadokoro <davidbtadokoro@usp.br>
Signed-off-by: David Tadokoro <davidbtadokoro@usp.br>